solo falta leer el resultado de post z hacer el insert en facturas

+++ src/main/java/gio/co/hospital/ws/ClienteSeguro/GetCliente.java
@@ -44,7 +44,7 @@ public class GetCliente extends HttpServlet {
             String url;
             String parClienteId = request.getParameter("dpi");
             if ((parClienteId != null) && !(parClienteId.equals(""))) {
-                int cId = Integer.parseInt(request.getParameter("dpi"));
+                long cId = Long.parseLong(request.getParameter("dpi"));
                 url = "http://25.65.236.60:8080/proyectoDB2-seguro/restC/cliente/getCliente?dpi=" + cId;
                 //url = "http://localhost:8080/proyectoDB2-seguro/restC/cliente/getCliente?dpi=" + cId;
             } else {
@@ -78,6 +78,7 @@ public class GetCliente extends HttpServlet {
     }
     
 }
+
